Revert "Course content refactor"
