+++ convex/_generated/api.d.ts
@@ -14,7 +14,6 @@ import type {
   FilterApi,
   FunctionReference,
 } from "convex/server";
-import type * as articles from "../articles";
 import type * as courseModules from "../courseModules";
 import type * as courses from "../courses";
 import type * as lib_authorization from "../lib/authorization";
@@ -34,7 +33,6 @@ import type * as users from "../users";
  * ```
  */
 declare const fullApi: ApiFromModules<{
-  articles: typeof articles;
   courseModules: typeof courseModules;
   courses: typeof courses;
   "lib/authorization": typeof lib_authorization;
