+++ convex/moduleSections.ts
@@ -9,7 +9,6 @@ import { asyncMap, getManyFrom } from "./lib/relationships";
 
 /* ADMIN ONLY
 ======================================= */
-const defaultSectionTitle = "Untitled section";
 
 export const findById = query({
   args: {
@@ -21,6 +20,7 @@ export const findById = query({
   },
 });
 
+const defaultSectionTitle = "Untitled section";
 export const create = mutation({
   args: {
     moduleId: v.id("modules"),
@@ -40,6 +40,7 @@ export const create = mutation({
       moduleId,
       type,
       title: defaultSectionTitle,
+      content: "",
       order: existingModuleSections.length + 1,
     });
   },
@@ -50,15 +51,17 @@ export const update = mutation({
     id: v.id("moduleSections"),
     type: v.optional(v.string()),
     title: v.optional(v.string()),
+    content: v.optional(v.string()),
     order: v.optional(v.number()),
   },
-  handler: async (ctx, { id, type, title, order }) => {
+  handler: async (ctx, { id, type, title, content, order }) => {
     await validateIdentity(ctx, { requireAdminRole: true });
 
     const existingSection = await ctx.db.get(id);
     await ctx.db.patch(id, {
       type: type || existingSection?.type || "text",
       title: title || existingSection?.title || defaultSectionTitle,
+      content: content ?? existingSection?.content,
       order: order ?? existingSection?.order,
     });
     return await ctx.db.get(id);
