+++ convex/schema.ts
@@ -6,7 +6,6 @@ export default defineSchema({
     name: v.string(),
     tokenIdentifier: v.string(),
     isAdmin: v.optional(v.boolean()),
-    roles: v.optional(v.array(v.string())),
   }).index("by_token", ["tokenIdentifier"]),
   tracking: defineTable({
     event: v.string(),
@@ -37,18 +36,23 @@ export default defineSchema({
     moduleId: v.id("modules"),
     type: v.string(),
     title: v.string(),
+    content: v.string(),
     order: v.number(),
-    contentId: v.optional(
+    contentBlockIds: v.optional(
       v.union(v.id("article"), v.id("video"), v.id("assessment"))
     ),
   })
     .index("by_moduleId", ["moduleId"])
     .index("by_order", ["order"]),
-  articles: defineTable({
+  article: defineTable({
     title: v.string(),
     content: v.string(),
   }),
-  assessments: defineTable({
+  video: defineTable({
+    title: v.string(),
+    url: v.string(),
+  }),
+  assessment: defineTable({
     title: v.string(),
     questions: v.array(
       v.object({
