+++ src/app/admin/courses/[id]/page.tsx
@@ -6,21 +6,21 @@ import { Doc, Id } from "../../../../../convex/_generated/dataModel";
 import { PageContent, PageHeader } from "@/lib/layout";
 import {
   Button,
+  ContentReader,
   Text,
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-  FlexRow,
 } from "@/lib/ui";
 import { useEffect, useState } from "react";
+import { GripVertical } from "lucide-react";
 import {
   SortableList,
   SortableListItem,
 } from "@/lib/providers/SortableListProvider";
-import { QuickEditCourseModulesForm } from "../../../../lib/forms/admin/CourseModules/QuickEditCourseModulesForm";
-import { QuickEditCourseForm } from "../../../../lib/forms/admin/Courses/QuickEditCourseForm";
+import { QuickEditCourseModulesForm } from "./QuickEditCourseModulesForm";
 
 interface AdminCoursePageProps {
   params: { id: string };
@@ -33,18 +33,6 @@ export default function AdminCoursePage({ params }: AdminCoursePageProps) {
 
   const [selectedModuleId, setSelectedModuleId] =
     useState<Id<"modules"> | null>(null);
-
-  const [selectedModuleSectionId, setSelectedModuleSectionId] =
-    useState<Id<"moduleSections"> | null>(null);
-
-  const moduleData = useQuery(
-    api.modules.findById,
-    selectedModuleId !== null
-      ? {
-          id: selectedModuleId,
-        }
-      : "skip"
-  );
 
   useEffect(() => {
     if (!course?.modules?.[0]?._id || !!selectedModuleId) return;
@@ -66,13 +54,10 @@ export default function AdminCoursePage({ params }: AdminCoursePageProps) {
         ]}
       />
       <PageContent>
-        <FlexRow className="justify-between">
-          <div className="space-y-0.5">
-            <Text className="text-2xl font-semibold">{course.title}</Text>
-            <Text className="text-muted-foreground">{course.description}</Text>
-          </div>
-          <QuickEditCourseForm courseId={params.id as Id<"courses">} />
-        </FlexRow>
+        <div className="space-y-0.5">
+          <Text className="text-2xl font-semibold">{course.title}</Text>
+          <Text className="text-muted-foreground">{course.description}</Text>
+        </div>
         <hr />
         <div className="flex flex-col lg:flex-row">
           <aside className="lg:w-1/4 lg:pr-8">
@@ -91,23 +76,8 @@ export default function AdminCoursePage({ params }: AdminCoursePageProps) {
               />
             </div>
           </aside>
-          <aside className="lg:w-1/4 lg:pr-4">
-            <div className="flex flex-col gap-4 lg:sticky lg:top-0">
-              <div className="flex items-center justify-between">
-                <Text className="font-bold">Module Sections</Text>
-              </div>
-              {moduleData && (
-                <ModuleSectionNav
-                  moduleId={params.id as Id<"modules">}
-                  sections={moduleData.sections}
-                  selectedModuleSectionId={selectedModuleSectionId}
-                  setSelectedModuleSectionId={setSelectedModuleSectionId}
-                />
-              )}
-            </div>
-          </aside>
-          <div className="flex-1 lg:w-1/2 lg:pl-8">
-            {selectedModuleSectionId && <p>{selectedModuleSectionId}</p>}
+          <div className="flex-1 lg:w-3/4 lg:pl-8">
+            {selectedModuleId && <Module id={selectedModuleId} />}
           </div>
         </div>
       </PageContent>
@@ -188,6 +158,19 @@ function ModuleNav({
                       {module.order}. {module.title ?? "Untitled section"}
                     </div>
                   </Button>
+                  {selectedModuleId === module?._id && (
+                    <div className="pl-4">
+                      {module.sections.map((section) => (
+                        <Text
+                          key={section._id}
+                          size="sm"
+                          className="py-2 truncate"
+                        >
+                          {section.title}
+                        </Text>
+                      ))}
+                    </div>
+                  )}
                 </div>
               </SortableListItem>
             ))}
@@ -215,96 +198,6 @@ function ModuleNav({
   );
 }
 
-function ModuleSectionNav({
-  moduleId,
-  sections,
-  selectedModuleSectionId,
-  setSelectedModuleSectionId,
-}: {
-  moduleId: Id<"modules">;
-  sections: Doc<"moduleSections">[];
-  selectedModuleSectionId: Id<"moduleSections"> | null;
-  setSelectedModuleSectionId: (id: Id<"moduleSections">) => void;
-}) {
-  const sortItems = sections.map((section) => section._id);
-
-  const updateSectionsOrder = useMutation(
-    api.moduleSections.updateOrder
-  ).withOptimisticUpdate((localStore, args) => {
-    const { idsInOrder } = args;
-    const updatedModuleSections = sections
-      .map((section, index) => ({
-        ...section,
-        order: idsInOrder.indexOf(section._id) + 1,
-      }))
-      .sort((a, b) => a.order - b.order);
-    const currentValue = localStore.getQuery(api.modules.findById, {
-      id: moduleId,
-    });
-    if (currentValue !== undefined) {
-      localStore.setQuery(
-        api.modules.findById,
-        {
-          id: moduleId,
-        },
-        { ...currentValue, sections: updatedModuleSections }
-      );
-    }
-  });
-
-  function handleOnUpdate(updatedItems: string[]) {
-    updateSectionsOrder({
-      idsInOrder: updatedItems as Id<"moduleSections">[],
-    });
-  }
-
-  return (
-    <>
-      <SortableList items={sortItems} onUpdate={handleOnUpdate}>
-        <div className="flex flex-col gap-2">
-          {sections.map((section) => (
-            <SortableListItem key={section._id} id={section._id}>
-              <Button
-                key={section?._id}
-                variant={
-                  selectedModuleSectionId === section?._id
-                    ? "secondary"
-                    : "ghost"
-                }
-                onClick={() => setSelectedModuleSectionId(section?._id)}
-                className="flex-1 truncate"
-              >
-                <div className="w-full text-left truncate">
-                  {section.order}. {section?.title ?? "Untitled section"}
-                </div>
-              </Button>
-            </SortableListItem>
-          ))}
-        </div>
-      </SortableList>
-      <div className="block lg:hidden pb-6">
-        <Select
-          onValueChange={(val) =>
-            setSelectedModuleSectionId(val as Id<"moduleSections">)
-          }
-          value={selectedModuleSectionId as string}
-        >
-          <SelectTrigger className="w-full">
-            <SelectValue placeholder="Select a module" />
-          </SelectTrigger>
-          <SelectContent>
-            {sections.map((section) => (
-              <SelectItem value={section._id} key={section._id}>
-                {section.title}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-      </div>
-    </>
-  );
-}
-
 function Module({ id }: { id: Id<"modules"> }) {
   const courseModule = useQuery(api.modules.findById, {
     id,
@@ -320,11 +213,7 @@ function Module({ id }: { id: Id<"modules"> }) {
       {courseModule.sections?.map((section) => (
         <div key={section._id}>
           <Text className="text-2xl pb-8">{section.title}</Text>
-          {courseModule.sections?.map((section) => (
-            <div key={section._id}>
-              <Text className="text-xl pb-4">{section.title}</Text>
-            </div>
-          ))}
+          <ContentReader content={section.content} />
           <hr className="my-8" />
         </div>
       ))}
