+++ src/app/admin/courses/lib/providers/CoursesProvider.ts
@@ -1,0 +1,5 @@
+// This file provides data to course components
+// It has a loading component that is displayed while data is being fetched
+// It doesn't render any component
+// It takes in children
+// Children can use the data provided by this component
