+++ src/app/admin/modules/QuickEditModuleForm.tsx
@@ -2,14 +2,14 @@
 
 import * as Yup from "yup";
 import { useMutation, useQuery } from "convex/react";
-import { api } from "../../../../../convex/_generated/api";
+import { api } from "../../../../convex/_generated/api";
 import {
   AdminQuickForm,
   AdminFormConfig,
   AdminFieldtype,
 } from "../AdminQuickForm";
 import { useToast } from "@/lib/hooks/useToast";
-import { Id } from "../../../../../convex/_generated/dataModel";
+import { Id } from "../../../../convex/_generated/dataModel";
 import { Pencil } from "lucide-react";
 import { Button } from "@/lib/ui";
