+++ src/app/admin/modules/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { PageContent, PageHeader } from "@/lib/layout";
-import { CreateModuleForm } from "../../../lib/forms/admin/Modules/CreateModuleForm";
-import { ModulesTable } from "../../../lib/views/admin/Modules/ModulesTable";
+import { CreateModuleForm } from "./CreateModuleForm";
+import { ModulesForm } from "./ModulesTable";
 import { Text } from "@/lib/ui";
 
 export default function AdminModulesPage() {
@@ -18,7 +18,7 @@ export default function AdminModulesPage() {
       <PageContent>
         <Text className="text-2xl font-semibold">Modules</Text>
         <hr />
-        <ModulesTable />
+        <ModulesForm />
       </PageContent>
     </>
   );
