+++ src/app/admin/page.tsx
@@ -1,7 +1,7 @@
 import { PageContent, PageHeader } from "@/lib/layout";
 import { Text } from "@/lib/ui";
 
-export default function AdminPage() {
+export default function AdminCoursesPage() {
   return (
     <>
       <PageHeader breadcrumbs={[{ href: "/admin", label: "Admin" }]} />
