--- src/lib/forms/admin/Articles/DeleteArticleButton.tsx
@@ -1,54 +1,0 @@
-import { useMutation } from "convex/react";
-import { api } from "../../../../../convex/_generated/api";
-import { Id } from "../../../../../convex/_generated/dataModel";
-import {
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-  Button,
-} from "@/lib/ui";
-import { Trash2 } from "lucide-react";
-
-export function DeleteArticleButton({
-  articleId,
-}: {
-  articleId: Id<"articles">;
-}) {
-  const deleteArticleSection = useMutation(api.articles.deleteById);
-
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger asChild>
-        <Button variant="ghost">
-          <Trash2 className="h-4 w-4 text-destructive" />
-        </Button>
-      </AlertDialogTrigger>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-          <AlertDialogDescription>
-            This action cannot be undone. This will permanently delete this
-            article in all courses and modules that use it.
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>Cancel</AlertDialogCancel>
-          <AlertDialogAction asChild>
-            <Button
-              variant="destructive"
-              onClick={() => deleteArticleSection({ id: articleId })}
-            >
-              Yes, I&lsquo;m sure.
-            </Button>
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
